LUN-674: Cleanup services for web
