+++ web/src/components/CategoryEntry.tsx
@@ -3,7 +3,7 @@ import { CategoryModel } from 'api-client'
 import iconPlaceholder from '../assets/IconPlaceholder.svg'
 import styled, { withTheme } from 'styled-components'
 import Highlighter from 'react-highlight-words'
-import normalizeSearchString from '../services/normalizeSearchString'
+import { normalizeSearchString } from '../utils/stringUtils'
 import type { ThemeType } from 'build-configs'
 import ContentMatcher from './ContentMatcher'
 import { Link } from 'react-router-dom'
