+++ web/src/components/EventListItem.tsx
@@ -4,7 +4,7 @@ import ListItem from './ListItem'
 import EventPlaceholder1 from '../assets/EventPlaceholder1.jpg'
 import EventPlaceholder2 from '../assets/EventPlaceholder2.jpg'
 import EventPlaceholder3 from '../assets/EventPlaceholder3.jpg'
-import textTruncator from '../services/textTruncator'
+import { textTruncator } from '../utils/stringUtils'
 
 export const NUM_OF_WORDS_ALLOWED = 15
