+++ web/src/components/NewsListItem.tsx
@@ -5,7 +5,7 @@ import CleanLink from './CleanLink'
 import LastUpdateInfo from './LastUpdateInfo'
 import { Description } from './ListItem'
 import { TFunction } from 'react-i18next'
-import textTruncator from '../services/textTruncator'
+import { textTruncator } from '../utils/stringUtils'
 import { DateFormatter, LOCAL_NEWS_TYPE, NewsType } from 'api-client'
 
 export const NUM_OF_WORDS_ALLOWED = 30
