+++ web/src/components/__tests__/EventListItem.spec.tsx
@@ -5,7 +5,7 @@ import EventPlaceholder1 from '../../assets/EventPlaceholder1.jpg'
 import EventPlaceholder2 from '../../assets/EventPlaceholder2.jpg'
 import EventPlaceholder3 from '../../assets/EventPlaceholder3.jpg'
 import EventListItem, { NUM_OF_WORDS_ALLOWED } from '../EventListItem'
-import textTruncator from '../../services/textTruncator'
+import { textTruncator } from '../../utils/stringUtils'
 import { ThemeProvider } from 'styled-components'
 import buildConfig from '../../constants/buildConfig'
 import { renderWithRouter } from '../../testing/render'
