+++ web/src/routes/EventsPage.tsx
@@ -20,7 +20,7 @@ import { createPath } from './index'
 import FailureSwitcher from '../components/FailureSwitcher'
 import Page, { THUMBNAIL_WIDTH } from '../components/Page'
 import PageDetail from '../components/PageDetail'
-import featuredImageToSrcSet from '../services/featuredImageToSrcSet'
+import featuredImageToSrcSet from '../utils/featuredImageToSrcSet'
 import { useTranslation } from 'react-i18next'
 import List from '../components/List'
 import Caption from '../components/Caption'
