--- web/src/services/normalizeSearchString.ts
@@ -1,5 +1,0 @@
-import normalizeStrings from 'normalize-strings'
-
-const normalizeSearchString = (str: string): string => normalizeStrings(str).toLowerCase()
-
-export default normalizeSearchString
