--- web/src/services/urlFromPath.ts
@@ -1,9 +1,0 @@
-/**
- * @param path: a slash-prefixed path
- * @returns {string} the url of the current host and the specified path
- */
-const urlFromPath = (path: string): string => {
-  return `${location.origin}${path}`
-}
-
-export default urlFromPath
