+++ web/src/utils/BrowserLanguageDetector.ts
