+++ web/src/utils/featuredImageToSrcSet.ts
