+++ web/src/utils/initSentry.ts
